requirements.txt
# pipreqs /home/jair/Proyectos/Patient_Churn_Prediction_Dataset_for_Healthcare --force

matplotlib==3.10.8
numpy==2.4.2
pandas==3.0.0
scikit_learn==1.8.0
seaborn==0.13.2
imbalanced-learn==0.10.1
PyYAML==6.0
xgboost==1.7.4